+++ packages/codemod/test/transforms/v1/fixtures/datasource-mssql/datasource-mssql.input.ts
@@ -4,4 +4,8 @@ const dataSource = new DataSource({
     type: "mssql",
     domain: "MYDOMAIN",
     username: "user",
+    options: {
+        isolation: "READ_COMMITTED",
+        connectionIsolationLevel: "REPEATABLE_READ",
+    },
 })
